backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.1
pydantic>=2.9.0
python-multipart==0.0.9
email-validator==2.2.0
slowapi==0.1.9
